refactor(mod-queue): move mod queue button out of the route view
board-buttons (reusable board chrome) imported ModQueueButton from the mod
queue route module, coupling a component to a route view and its private
helpers and CSS. Move the button to src/components/mod-queue-button/ and pull
the shared pieces it needs out of the route:

- useLocallyModeratedModQueueFeed -> src/hooks/use-locally-moderated-mod-queue-feed.ts
- ModQueueCommunityMetadataLoader -> its own component (used by page and button)
- getAddressListKey/getAddressListFromKey -> src/lib/utils/mod-queue-utils.ts

The moved code is byte-identical to the originals. The button's count-badge
CSS is copied into its own module (the route keeps its copy for the page's
board count display, which shares those classes), so appearance is unchanged.
board-buttons now imports from the component, and the route no longer exports
the button. Reusable components no longer depend on a route view.

Implements plans/003-move-mod-queue-button.md.
