+++ src/components/mod-queue-button/index.ts
@@ -1,0 +1,1 @@
+export { ModQueueButton } from './mod-queue-button';
