+++ src/components/mod-queue-community-metadata-loader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './mod-queue-community-metadata-loader';
